refactor(visual): cleanup + consolidate shared logic
Cleanup (per code review on PR #123):
- Drop empty `{logName}` placeholder from `formatImageName` so baseline
  filenames lose the literal double-dash (`home-200x272.png` instead
  of `home--200x272.png`).
- Remove cargo-culted Appium-only `blockOutSideBar/StatusBar/ToolBar`
  flags from the visual service config — they have no effect on
  desktop captures.
- Document why `autoSaveBaseline: true` is correct for the spike's
  ephemeral CI runners and what downstream users would change for
  real CI strictness.
- Update VRT-SPIKE-FINDINGS.md to reflect green CI matrix run 2.

Consolidation:
- Add a workspace-root `wdio.visual.shared.ts` exporting
  `asVisualConfig(base)`. The 6 `wdio.*.visual.conf.ts` files now
  delegate to it, dropping each from 8 lines to 3.
- Add per-package `test/lib/visual.ts` exporting `MAX_MISMATCH_PCT`
  + `stabilise()`. Each visual.spec.ts now imports from
  `../lib/visual.ts` (mirrors the existing `lib/utils.ts` pattern in
  each package), removing the inline helper duplication.

Net: ~50 lines saved across the PR. Each package still owns its
visual config + spec; only genuinely-shared logic is hoisted.

Both consolidations smoke-tested locally (electron-builder +
tauri-embedded) — match path passes after baseline-write.

Also: drop `.md` from the lint-staged biome-format glob. Biome 2.x
doesn't format markdown, so the previous attempt to handle the
"only-md staged" case via `--files-ignore-unknown=true` still failed
because biome errors when zero files are processed even with the
flag. Removing `.md` from the glob is the simplest correct fix.
