+++ package.json
@@ -83,7 +83,7 @@
   "lint-staged": {
     "**/*.{js,ts}": ["biome check --write --formatter-enabled=false --linter-enabled=true"],
     "packages/*/test/**/*.{js,ts}": ["eslint --fix --cache"],
-    "**/*.{json,md}": ["biome format --write --files-ignore-unknown=true"]
+    "**/*.json": ["biome format --write"]
   },
   "packageManager": "pnpm@10.3.0+sha512.ee592eda8815a8a293c206bb0917c4bb0ff274c50def7cbc17be05ec641fc2d1b02490ce660061356bd0d126a4d7eb2ec8830e6959fb8a447571c631d5a2442d",
   "pnpm": {
