+++ packages/electron-forge/wdio.base.conf.ts
@@ -76,11 +76,13 @@ export const visualService: Services.ServiceEntry = [
   {
     baselineFolder: join(__dirname, '__visual__', process.platform, process.arch, 'baseline'),
     screenshotPath: join(__dirname, '__visual__', process.platform, process.arch, 'actual'),
-    formatImageName: '{tag}-{logName}-{width}x{height}',
+    formatImageName: '{tag}-{width}x{height}',
+    // CI runners are ephemeral, so auto-saving the baseline on first run keeps
+    // the matrix self-contained (each job writes once, then validates the
+    // match path on the second invocation in `test:visual`). For a real
+    // downstream project you'd typically want `!process.env.CI` so missing
+    // baselines fail loudly in CI and only update via an explicit flow.
     autoSaveBaseline: true,
-    blockOutSideBar: true,
-    blockOutStatusBar: true,
-    blockOutToolBar: true,
   },
 ];
