+++ packages/electron-script/wdio.visual.conf.ts
@@ -1,8 +1,4 @@
-import type { Options } from '@wdio/types';
+import { asVisualConfig } from '../../wdio.visual.shared.ts';
 import { config as base } from './wdio.conf.ts';
 
-export const config: Options.Testrunner = {
-  ...base,
-  specs: ['./test/visual/**/*.spec.ts'],
-  exclude: [],
-};
+export const config = asVisualConfig(base);
